+++ src/main/java/com/shy_polarbear/server/global/auth/security/PrincipalDetailService.java
@@ -1,4 +1,4 @@
-package com.shy_polarbear.server.global.config.security;
+package com.shy_polarbear.server.global.auth.security;
 
 import com.shy_polarbear.server.domain.user.model.User;
 import com.shy_polarbear.server.domain.user.repository.UserRepository;
