+++ src/main/java/com/shy_polarbear/server/global/auth/security/SecurityConfig.java
@@ -1,10 +1,10 @@
-package com.shy_polarbear.server.global.config.security;
+package com.shy_polarbear.server.global.auth.security;
 
 
 
-import com.shy_polarbear.server.global.config.jwt.JwtAuthenticationEntryPoint;
-import com.shy_polarbear.server.global.config.jwt.JwtAuthenticationFilter;
-import com.shy_polarbear.server.global.config.jwt.JwtProvider;
+import com.shy_polarbear.server.global.auth.jwt.JwtProvider;
+import com.shy_polarbear.server.global.auth.jwt.JwtAuthenticationEntryPoint;
+import com.shy_polarbear.server.global.auth.jwt.JwtAuthenticationFilter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
